Extract IssueProject and Issue from HostedRepository and PullRequest
